+++ src/test/java/com/example/kinnibackend/category/ProductFilteringControllerTest.java
@@ -14,7 +14,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @SpringBootTest
 @AutoConfigureMockMvc
 @TestInstance(TestInstance.Lifecycle.PER_METHOD)
-public class CategoryControllerTest {
+public class ProductFilteringControllerTest {
 
     @Autowired
     private MockMvc mockMvc;
